pyproject.toml
[project]
name = "dbx-sql-runner"
version = "0.1.0"
description = "A lightweight SQL transformation tool for Databricks SQL"
authors = [{ name = "Sharma", email = "munish7771@gmail.com" }]
dependencies = [
    "networkx",
    "databricks-sql-connector[pyarrow]",
    "PyYAML"
]
readme = "README.md"

[project.scripts]
dbx-sql-runner = "dbx_sql_runner.cli:main"

[tool.setuptools]
packages = ["dbx_sql_runner"]